requirements.txt
Click==7.0
Flask==1.0.2
gunicorn==19.9.0
ItsDangerous==1.0.0
Jinja2==2.10
MarkupSafe==1.0
mysql-connector==2.1.6
mysql-connector-python==8.0.13
numpy==1.14.3
Pillow==5.1.0
protobuf==3.6.1
Werkzeug==0.14.1
pandas==0.23.4